Rename machine_charm.py to charm.py
Needed for upcoming charmcraft 3 migration because of https://github.com/canonical/charmcraft/issues/1994
