+++ charmcraft.yaml
@@ -67,4 +67,4 @@ parts:
     # TODO: enable after https://github.com/canonical/charmcraft/issues/1456 fixed
     charm-strict-dependencies: false
     charm-requirements: [requirements.txt]
-    charm-entrypoint: src/machine_charm.py
+    charm-entrypoint: src/charm.py
