+++ tests/unit/scenario_/database_relations/test_database_relations_breaking.py
@@ -7,7 +7,7 @@ import ops
 import pytest
 import scenario
 
-import machine_charm
+import charm
 
 from . import combinations
 
@@ -18,7 +18,7 @@ def output_state(
     event: scenario.Event,
     secrets: list[scenario.Secret] = [],
 ) -> scenario.State:
-    context = scenario.Context(machine_charm.MachineSubordinateRouterCharm)
+    context = scenario.Context(charm.MachineSubordinateRouterCharm)
     input_state = scenario.State(
         relations=[*relations, scenario.PeerRelation(endpoint="upgrade-version-a")],
         secrets=secrets,
